REFACTOR: Using server actions instead of endpoints - Changed pages into server component pages - Importing client side components into server side component pages
