+++ app/(main)/chats/[id]/page.tsx
@@ -1,13 +1,7 @@
-"use client";
-
-import { useEffect, useRef, useState } from "react";
-import { useParams, useSearchParams } from "next/navigation";
-import { apiGetChat } from "@/lib/chat/api";
-import type { ChatWithMessages } from "@/lib/chat/api";
-import { useAppStore } from "@/lib/store";
+import { getChat } from "@/lib/actions/chats/get-chat";
+import { ChatPageClient } from "./_components/chat-page-client";
 import type { Chat, Message } from "@/lib/store";
-import { ChatUI } from "@/components/chat/chat-ui";
-import { Skeleton } from "@/components/ui/skeleton";
+import type { ChatWithMessages } from "@/lib/actions/chats/types";
 
 function buildChatFromRows(row: ChatWithMessages): Chat {
   const messages: Record<string, Message> = {};
@@ -40,55 +34,25 @@ function buildChatFromRows(row: ChatWithMessages): Chat {
   };
 }
 
-/**
- * Single chat view page that renders ChatUI for the chat matching the URL parameter.
- * Fetches the chat and its messages from the DB on mount, hydrates Zustand, then renders.
- * Route: /chats/[id].
- *
- * @author Maruf Bepary
- */
-export default function ChatPage() {
-  const params = useParams();
-  const searchParams = useSearchParams();
-  const chatId = params.id as string;
-  const upsertChat = useAppStore((state) => state.upsertChat);
-  const [isLoading, setIsLoading] = useState(true);
-  const [notFound, setNotFound] = useState(false);
-  const initialMessage = searchParams.get("msg") ?? undefined;
-  const didSendInitial = useRef(false);
+export default async function ChatPage({
+  params,
+  searchParams,
+}: {
+  params: Promise<{ id: string }>;
+  searchParams: Promise<{ msg?: string }>;
+}) {
+  const { id: chatId } = await params;
+  const { msg } = await searchParams;
 
-  useEffect(() => {
-    apiGetChat(chatId)
-      .then((data) => {
-        const chat = buildChatFromRows(data);
-        upsertChat(chat);
-      })
-      .catch(() => {
-        setNotFound(true);
-      })
-      .finally(() => {
-        setIsLoading(false);
-      });
-  }, [chatId, upsertChat]);
-
-  if (isLoading)
-    return (
-      <div className="flex-1 p-6">
-        <Skeleton className="h-8 w-48" />
-      </div>
-    );
-  if (notFound)
+  let chat: Chat;
+  try {
+    const data = await getChat(chatId);
+    chat = buildChatFromRows(data);
+  } catch {
     return (
       <div className="flex-1 p-6 text-muted-foreground">Chat not found.</div>
     );
+  }
 
-  return (
-    <ChatUI
-      chatId={chatId}
-      initialMessage={!didSendInitial.current ? initialMessage : undefined}
-      onInitialMessageSent={() => {
-        didSendInitial.current = true;
-      }}
-    />
-  );
+  return <ChatPageClient initialChat={chat} initialMessage={msg} />;
 }
