+++ lib/actions/chats/persist-message.ts
@@ -1,31 +1,36 @@
+"use server";
+
 import { auth } from "@/lib/auth/auth";
 import { db } from "@/drizzle/db";
 import { chat, message } from "@/drizzle/schema";
 import { eq, and } from "drizzle-orm";
 import { headers } from "next/headers";
+import type { MessageRow } from "./types";
 
-export async function POST(
-  req: Request,
-  { params }: { params: Promise<{ chatId: string }> },
-) {
+export async function persistMessage(
+  chatId: string,
+  msg: { id: string; role: string; content: string; parentId: string | null },
+): Promise<MessageRow> {
   const session = await auth.api.getSession({ headers: await headers() });
-  if (!session) return new Response("Unauthorized", { status: 401 });
-
-  const { chatId } = await params;
+  if (!session) throw new Error("Unauthorized");
 
   const [chatRow] = await db
     .select({ id: chat.id })
     .from(chat)
     .where(and(eq(chat.id, chatId), eq(chat.userId, session.user.id)));
 
-  if (!chatRow) return new Response("Not Found", { status: 404 });
-
-  const { id, role, content, parentId } = await req.json();
+  if (!chatRow) throw new Error("Not Found");
 
   const [newMessage] = await db
     .insert(message)
-    .values({ id, chatId, role, content, parentId: parentId ?? null })
+    .values({
+      id: msg.id,
+      chatId,
+      role: msg.role,
+      content: msg.content,
+      parentId: msg.parentId ?? null,
+    })
     .returning();
 
-  return Response.json(newMessage, { status: 201 });
+  return newMessage as MessageRow;
 }
